feat: 添加Windows平台支持并改进代码安全性
- 将sprintf替换为sprintf_s以提高安全性
- 为Windows平台添加本地时间处理支持
- 实现简化的CLI参数解析器以兼容Windows
- 添加最小构建模式支持
- 改进测试输出格式和错误处理
- 添加构建脚本(build.sh/build.bat)

+++ src/bd_scanner.h
@@ -15,7 +15,7 @@ struct BDStructure {
 struct PlaylistInfo {
     std::string path;          // Path to the playlist file
     uint64_t duration;         // Duration in milliseconds
-    int segmentCount;          // Number of segments
+    size_t segmentCount;       // Number of segments
 };
 
 /**
